(#12,pgnv) Patch chessground/render.js on pgnv build
When board are rendered in MW, no pieces are displayed.
When comparing HTML (through browser inspection) with the one produced
with example.hmtl, we see the following differences.

In example.html, we have:

    <piece class="black rook" style="transform: translate(0px, 0px);"></piece>

In MW HTML, we have:

    <piece class="blackrook" style="transform: translate(0px, 0px);"></piece>

Strangely enough, we miss a SPACE character.

Narrowing down the issue in pgnv.js leads us to chessboard module,
file render.js.

This patch is an ugly fix to that issue. It changes

    function pieceNameOf(piece) {
        return `${piece.color} ${piece.role}`;
    }

into

    function pieceNameOf(piece) {
        return `${piece.color}` + " " + `${piece.role}`;
    }

Why the original code does not work is a mystery.

See https://github.com/ornicar/chessground/issues/169